Added fonts
This might be able to add the Libre Baskerville font to all variant SVGs without embedding the fonts in the SVGs. That would allow the new style maps to render more nicely, but at the same time make them dependent on the web view (e.g. diplicity android wouldn't render)