Probably the ugliest refactor ever
I made the orders not take up so much screen size, and be still informative.

We need to make them appear after a certain delay (need to play with it to feel what is good), because now they appear too often. But not too long so that it's not possible to find this by accident.