Local Setting for Beta
Added a local setting.
If no setting, don't show (hide by default).
If setting is true, show

In settings, it's writing out clearly, but then things start becoming a mess.

I used prettier and it's still the WRONG prettier, so if you help me on this you need to re-prettier it before you start working on it. I added TODOs  where you can find the code. We need to find a solution for that as well.

I'm SO happy when this damn feature is out! So close...