MoveViaConvoy now doesn't crash when not resolve
During the current phase, when an MVC order is given the game doesn't crash.

However, it doesn't render the move order correctly - likely because the MVC order is not 'read' correctly.

Might need a simple "drawArrow" element in the MVC in dippymap - but then we need to include the 'isresolved' element